添加博客

+++ src/main/java/test/spring/dao/AccountDAO.java
@@ -8,15 +8,15 @@ import org.springframework.stereotype.Repository;
 @Repository
 public class AccountDAO extends AbstractDAO {
 	
-	public boolean createAccount(int userId) {
+	public boolean createAccount(long userId) {
 		return sqlSession.insert("AccountDAO.createAccount", userId) > 0;
 	}
 	
-	public double getBalance(int userId) {
+	public double getBalance(long userId) {
 		return sqlSession.selectOne("AccountDAO.getBalance", userId);
 	}
 	
-	public boolean modifyAmmount(int userId, double ammount) {
+	public boolean modifyAmmount(long userId, double ammount) {
 		Map<String, Object> params = new HashMap<String, Object>();
 		params.put("userId", userId);
 		params.put("ammount", ammount);
