+++ src/main/java/test/spring/web/WebUtils.java
@@ -37,10 +37,10 @@ public class WebUtils extends org.springframework.web.util.WebUtils {
 		response.addCookie(cookie);
 	}
 	
-	public static int getLoginUserId(HttpServletRequest request) {
+	public static long getLoginUserId(HttpServletRequest request) {
 		Cookie cookie = WebUtils.getCookie(request, "uid");
 		if(cookie != null) {
-			return Integer.parseInt(cookie.getValue());
+			return Long.parseLong(cookie.getValue());
 		}
 		
 		return -1;
